+++ src/main/scala/utils/GridConversion.scala
@@ -69,9 +69,17 @@ object GridConversion {
       subnet: Int = 1,
       vTarget: ComparableQuantity[Dimensionless] = 1d.asPu
   ): NodeInput = {
+    val id_string =
+      if (id.isEmpty) {
+        val coordinatesString = coordinate.getCoordinates.mkString
+        val cleanedString =
+          coordinatesString.replace(",", "").replace("NaN", "").trim
+        "Node " + cleanedString
+      } else id
+
     new NodeInput(
       UUID.randomUUID(),
-      id,
+      id_string,
       vTarget,
       isSlack,
       coordinate,